recommendation_ai/requirements.txt
# PostgreSQL 数据库连接
psycopg2==2.9.9          # 支持 Python 3.10+

# 数据处理
pandas==1.5.3            # pandas 1.5.x 完美支持 Python 3.10
numpy==1.23.5            # 推荐搭配 pandas 使用，推荐加上

# 推荐系统库
scikit-surprise==1.1.3   # 支持 Python 3.10（注意不能升级到 scikit-surprise 2.x）
psycopg2-binary>=2.9

# 测试工具
pytest==7.3.1            # 支持 Python 3.10
